app/database.py
import os
from dotenv import load_dotenv
import sqlalchemy as _sql
import sqlalchemy.orm as _orm
import sqlalchemy.ext.declarative as _declarative


load_dotenv(".env")

DATABASE_URL = os.environ["DATABASE_URL"]

engine = _sql.create_engine(DATABASE_URL)

SessionLocal = _orm.sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = _declarative.declarative_base()